net: Extract skb_data
